requirements.txt
torch
torchvision
pillow
scikit-learn
flask
requests